# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TVS Bot Admin",
    "summary": "Configure the TVS WhatsApp bot and inspect what it is doing.",
    "description": "Every endpoint except the health probe and the panel HTML requires\n`Authorization: Bearer <ADMIN_TOKEN>`. One shared token, no roles.\n\nIf `ADMIN_TOKEN` is unset on the server every admin endpoint returns **503** —\nthe panel is disabled, not open.\n\nTwo things to know before building a client:\n\n* **Conversations group by `mobile`, not `session`.** A returning customer gets\n  a fresh `session` id whenever their hour-long window lapses. Turns recorded\n  before the `mobile` column shipped have `mobile: \"\"` and cannot be found by\n  number; there is no backfill.\n* **`PUT /admin/api/config` replaces the whole document.** There is no partial\n  update and no version history, so read, edit, then write back.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/admin/api/meta": {
      "get": {
        "tags": [
          "configuration"
        ],
        "summary": "Allowed values and factory defaults",
        "operationId": "meta_admin_api_meta_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Meta Admin Api Meta Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin token."
          },
          "503": {
            "description": "Admin panel disabled (ADMIN_TOKEN unset)."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/admin/api/config": {
      "get": {
        "tags": [
          "configuration"
        ],
        "summary": "Read the live bot configuration",
        "operationId": "read_config_admin_api_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Read Config Admin Api Config Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin token."
          },
          "503": {
            "description": "Admin panel disabled (ADMIN_TOKEN unset)."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "configuration"
        ],
        "summary": "Replace the bot configuration",
        "operationId": "write_config_admin_api_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Write Config Admin Api Config Put"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin token."
          },
          "503": {
            "description": "Admin panel disabled (ADMIN_TOKEN unset)."
          },
          "400": {
            "description": "Validation failed; detail names the field."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/admin/api/status": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "Is the bot working?",
        "description": "Is the bot working? One verdict plus the six checks behind it.",
        "operationId": "operational_status_admin_api_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Operational Status Admin Api Status Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin token."
          },
          "503": {
            "description": "Admin panel disabled (ADMIN_TOKEN unset)."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/admin/api/session/reset": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Hard-reset one customer's session",
        "description": "Hard-reset one customer: their next message starts from scratch.\n\nDeduplication keys are deliberately untouched — they are keyed per\nmessage, not per customer, and any new inbound message carries a\nfresh id.",
        "operationId": "reset_session_admin_api_session_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Reset Session Admin Api Session Reset Post"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin token."
          },
          "503": {
            "description": "Redis unavailable."
          },
          "400": {
            "description": "mobile missing or blank."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/admin/api/interactions/export.csv": {
      "get": {
        "tags": [
          "conversations"
        ],
        "summary": "Download conversations as CSV",
        "operationId": "export_interactions_admin_api_interactions_export_csv_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mobile",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mobile"
            }
          },
          {
            "name": "session",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Channel"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "needs_review",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Needs Review"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin token."
          },
          "503": {
            "description": "Admin panel disabled (ADMIN_TOKEN unset)."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/api/leads/export.csv": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "Download all leads as CSV",
        "operationId": "export_leads_admin_api_leads_export_csv_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin token."
          },
          "503": {
            "description": "Admin panel disabled (ADMIN_TOKEN unset)."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "ADMIN_TOKEN",
        "scheme": "bearer"
      }
    }
  },
...
}